Update README with measure power draw.
Measured to use 0.25A idle (1.25W) and 0.5A full on (2.5W).
This is well below the maximum power draw of the LEDs and seems
consistent with the fact that we're keeping the LEDs really dim. I intend
to order at least a 10W power supply to give plenty of buffer for
tinkering with the software in the future.